separate auth and user
