+++ src/Controllers/UsersController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using SACA.Entities.Requests;
+using SACA.Constants;
 using SACA.Entities.Responses;
 using SACA.Extensions;
 using SACA.Interfaces;
@@ -8,20 +8,18 @@ using SACA.Repositories.Interfaces;
 
 namespace SACA.Controllers
 {
-    public class AuthController : BaseApiController, IApiMarker
+    public class UsersController : BaseApiController
     {
         private readonly IS3Service _s3Service;
         private readonly MapperlyMapper _mapper;
-        private readonly ICategoryRepository _categoryRepository;
         private readonly IUserRepository _userRepository;
         private readonly IUnityOfWork _uow;
         private readonly IUserService _userService;
 
-        public AuthController(
+        public UsersController(
             IUserService userService,
             IS3Service s3Service,
             MapperlyMapper mapper,
-            ICategoryRepository categoryRepository,
             IUserRepository userRepository,
             IUnityOfWork unityOfWork
         )
@@ -29,48 +27,25 @@ namespace SACA.Controllers
             _userService = userService;
             _s3Service = s3Service;
             _mapper = mapper;
-            _categoryRepository = categoryRepository;
             _userRepository = userRepository;
             _uow = unityOfWork;
         }
-
-        [AllowAnonymous]
-        [HttpPost("signin")]
-        public async Task<ActionResult<UserResponse>> SignIn(SignInRequest signInRequest)
-        {
-            try
-            {
-                return await _userService.LogInAsync(signInRequest.Email, signInRequest.Password, signInRequest.Remember);
-            }
-            catch (ArgumentException argumentException)
-            {
-                return BadRequest(new ProblemDetails { Title = nameof(BadRequest), Detail = argumentException.Message });
-            }
-        }
-
-        [AllowAnonymous]
-        [HttpPost("signup")]
-        public async Task<ActionResult<UserResponse>> Create(SignUpRequest signUpRequest)
-        {
-            try
-            {
-                var applicationUser = await _userService.SignUpAsync(signUpRequest);
-                await _userService.AssignRolesAsync(applicationUser, signUpRequest.Roles);
-                var userResponse = await _userService.LogInAsync(signUpRequest.Email, signUpRequest.Password);
-
-                var user = await _userRepository.GetUserAsync(userResponse.Id);
-                user.Categories = await _categoryRepository.GetCommonCategoriesAsync();
-
-                await _uow.SaveChangesAsync();
-
-                return userResponse;
-            }
-            catch (ArgumentException argumentException)
-            {
-                return BadRequest(new ProblemDetails { Title = nameof(BadRequest), Detail = argumentException.Message });
-            }
-        }
         
+        [Authorize(Roles = AuthorizationConstants.Roles.Superuser)]
+        [HttpGet]
+        public async Task<ActionResult<List<UserResponse>>> GetAll()
+        {
+            var usersResponse = await _userRepository.GetUsersAsync();
+
+            foreach (var userResponse in usersResponse)
+            {
+                var user = _mapper.MapToApplicationUser(userResponse);
+                userResponse.Roles = await _userService.GetRolesAsync(user);
+            }
+
+            return usersResponse;
+        }
+
         [AllowAnonymous]
         [HttpGet("{id}")]
         public async Task<ActionResult<UserResponse>> Get(int id)
@@ -93,8 +68,8 @@ namespace SACA.Controllers
                 return BadRequest(new ProblemDetails { Title = nameof(BadRequest), Detail = argumentException.Message });
             }
         }
-
-        [HttpDelete("user")]
+        
+        [HttpDelete]
         public async Task<ActionResult<UserResponse>> Remove()
         {
             var userId = User.GetId();
@@ -104,6 +79,24 @@ namespace SACA.Controllers
             }
 
             var user = await _userRepository.GetUserAsync(userId.Value);
+            if (user is null)
+            {
+                return BadRequest();
+            }
+
+            await _s3Service.RemoveFolderAsync(user.Id.ToString());
+            await _userService.DeleteAsync(user);
+
+            await _uow.SaveChangesAsync();
+
+            return _mapper.MapToUserResponse(user);
+        }
+
+        [Authorize(Roles = AuthorizationConstants.Roles.Superuser)]
+        [HttpDelete("{id}")]
+        public async Task<ActionResult<UserResponse>> Remove(int id)
+        {
+            var user = await _userRepository.GetUserAsync(id);
             if (user is null)
             {
                 return BadRequest();
