+++ tests/SACA.Tests.Integration/UsersController/DeleteUserAuthControllerTests.cs
@@ -1,11 +1,12 @@
 using FluentAssertions;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using SACA.Entities.Responses;
+using SACA.Tests.Integration.AuthController;
 using System.Net;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
 
-namespace SACA.Tests.Integration.AuthController;
+namespace SACA.Tests.Integration.UsersController;
 
 public class DeleteUserAuthControllerTests : IClassFixture<TestFactory>
 {
@@ -21,13 +22,13 @@ public class DeleteUserAuthControllerTests : IClassFixture<TestFactory>
     public async Task<UserResponse> Should_DeleteUser_WhenUserExist(int id)
     {
         // Arrange
-        var signInUserAuthControllerTests = new SignInUserAuthControllerTests(_testFactory);
+        var signInUserAuthControllerTests = new SignInAuthControllerTests(_testFactory);
 
         // Act
         var user = await signInUserAuthControllerTests.Should_SignIn_WhenUserExist(id);
         _client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(JwtBearerDefaults.AuthenticationScheme, user.Token);
 
-        var response = await _client.DeleteAsync("v2/Auth/user");
+        var response = await _client.DeleteAsync("v2/Users");
         var userResponse = await response.Content.ReadFromJsonAsync<UserResponse>();
 
         // Assert
